+++ src/Router.tsx
@@ -6,6 +6,7 @@ import MainLayout from '#/layout/MainLayout';
 import AuthPage from '#/page/Auth';
 import CreateAccountPage from '#/page/Auth/CreateAccount';
 import FeedPage from '#/page/Feed';
+import FeedCreatePage from '#/page/Feed/FeedCreate';
 import GroupConnectionPage from '#/page/group/GroupConnection';
 import HomePage from '#/page/Home';
 
@@ -35,6 +36,7 @@ const router = createBrowserRouter([
                 element: <HomePage />,
             },
             { path: PAGE_PATH.feed, element: <FeedPage /> },
+            { path: PAGE_PATH.createFeed, element: <FeedCreatePage /> },
         ],
     },
 ]);
